third_party/blink/web_tests/external/wpt/fetch/fetch-later/send-on-deactivate.tentative.https.window-expected.txt

This is a testharness.js-based test.
FAIL fetchLater() sends on page entering BFCache if BackgroundSync is off. assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
FAIL Call fetchLater() when BFCached with activationTimeout=0 sends immediately. assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
FAIL fetchLater() sends on navigating away a page w/o BFCache. assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
FAIL fetchLater() with backgroundTimeout=1m sends on page entering BFCache if BackgroundSync is off.  assert_equals: Number of sent beacons does not match expected count: expected 1 but got 0
Harness: the test ran to completion.
